+++ NEWS.md
@@ -1,5 +1,6 @@
 # Changes in v0.9.3
 
+* Add `textstat_context()` and `char_context()` to replace `char_keyness()`.
 * Make the absolute sum of seed weight equal to 1.0 in both upper and lower ends. 
 * `textplot_terms()` takes glob patterns in character vector or a dictionary object.
 * `char_keyness()` no longer raise error when no patter is found in tokens object.